Deduplicate code for final fold inside bridge_unindexed
Instead of having the final .fold_with() being coded in two places,
combine the control flow so that it happens in only one.

The motivation is that the .fold_with() call represents the "inner loop"
and it seems bad to potentially duplicate that and have it be inlined
twice and optimized in two different locations.